+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- backend support to B2BQUOTES-28 updating the expirationDate field
+
+
 ## [1.2.0] - 2022-02-18
 
 ### Added