+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2022-03-10
+
 ### Added
 
 - Scheduler running twice a day changing the quotes status to expired when it needs